CBD-456 Separate document body and meta arguments to map function
Sends non-JSON values as byte arrays for the 'doc' argument.
Numbers and ascii text can be decoded trivially from this capability,
extending our support for legacy applications.

The document JSON format on the wire for bulk docs operations
now looks like this:

    {
      "json":{"user":"data","_anything":"goes"},
      "meta":{"id":"mydocid"}
    }

Or like this:

    {
      "base64":"bXliYXNlNjRiaW5hcnlkYXRh",
      "meta":{"id":"mydocid"}
    }

There can be additional data in the metadata object. Specification is on
the Jira ticket.

This change off support for _id lookups in include_docs queries, although
with server level include docs going away as well, this change should
be of little consequence. Include_docs is  not removed as it is used
heavily by tests, in the future it will be renamed to x_include_docs
across both tests and query  parser, in order to reflect the non-
supported nature.

Ticket with the specifications documentation:
http://www.couchbase.com/issues/browse/CBD-456

Admin UI and ns_server changes are in this change:
http://review.couchbase.org/#change,18028

Depends on this change in testrunner for view tests:
http://review.couchbase.org/#change,17768

Squashed in an old patch: http://review.couchbase.org/#change,18111

TODO: EP engine doesn't need to check for _fields when  setting JSON
flag. This can be accomplished in a future patch.

Related issue (not addressed by this patch): views and _all_docs
output (as well as XDCR JSON format) needs to understand non-UTF8
keys: http://www.couchbase.com/issues/browse/CBD-453

Change-Id: Ic8e3c5be07bd59b08f618b6f735e46c96ee4c9f2